Update main.css
I managed-after 40+ hours of wok-to get the mismatch rate to 0%!!!! The last stretch was converting .text classes to the much simpler p class. Once I found the correct font size everything fell into place. I was dumbfounded.
I've triple checked the code and am satisfied that everything is in line with Udacity's Style Guide. CSS is in shorthand, all is lowercase, color codes that can be reduced have been. ul and .col-md-7 have padding shorthands with three 0px's, but this creates a 0% mismatch whereas any other way does not.